src/requirements.txt
google-genai
prompt_toolkit
python-dotenv